Inverse verbosity option
Quieting the program did never silence all wordy ffmpeg runs. This was
Because the last, actual encoding one could be running for a very long
Time, so passing it through provides status information to the user.
Additionally, it passes on the question whether to overwrite an existing
File.

A -q flag turns most programs completely silent. As this program merely
Switches from "talkative" to "wall of text", verbose describes the logic
Much more appropriate.