Support templates in output file name
This eases the work with complex filenames or aids the creation of
scripts utilizing hardsub where the input file may not be known prior.
Limited path changes are also possible.

As it makes little sense to encode multiple files to the same target
location, it also paves the way for batch processing.